[evals] Consolidate overlapping adoption cases
