--- evals/evals/agent-042-enable-ppr/app/ProductList.tsx
@@ -1,18 +1,0 @@
-async function getProducts() {
-  const res = await fetch('https://api.example.com/products')
-  return res.json()
-}
-
-export async function ProductList() {
-  const products = await getProducts()
-
-  return (
-    <ul>
-      {products.map((p: { id: string; name: string; price: number }) => (
-        <li key={p.id}>
-          {p.name} - ${p.price}
-        </li>
-      ))}
-    </ul>
-  )
-}
