+++ evals/evals/agent-042-enable-ppr/app/layout.tsx
@@ -1,18 +1,16 @@
-import type { Metadata } from 'next'
+import Link from 'next/link'
+import type { ReactNode } from 'react'
 
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode
-}>) {
+export default function RootLayout({ children }: { children: ReactNode }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body>
+        <nav>
+          <Link href="/news">Newsroom</Link>{' '}
+          <Link href="/preview">Editor preview</Link>
+        </nav>
+        {children}
+      </body>
     </html>
   )
 }
