+++ evals/evals/agent-042-enable-ppr/next.config.ts
@@ -1,5 +1,10 @@
 import type { NextConfig } from 'next'
 
-const nextConfig: NextConfig = {}
+const nextConfig: NextConfig = {
+  experimental: {
+    // @ts-expect-error -- legacy config retained for the migration exercise
+    dynamicIO: true,
+  },
+}
 
 export default nextConfig
