--- evals/evals/agent-048-cache-components-config-semantics/app/page.tsx
@@ -1,3 +1,0 @@
-export default function HomePage() {
-  return <h1>Northstar Publishing</h1>
-}
