--- evals/evals/agent-048-cache-components-config-semantics/next.config.ts
@@ -1,10 +1,0 @@
-import type { NextConfig } from 'next'
-
-const nextConfig: NextConfig = {
-  experimental: {
-    // @ts-expect-error -- legacy config retained for the migration exercise
-    dynamicIO: true,
-  },
-}
-
-export default nextConfig
