--- evals/evals/agent-048-cache-components-config-semantics/package.json
@@ -1,23 +1,0 @@
-{
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "next": "^16",
-    "react": "19.1.0",
-    "react-dom": "19.1.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "typescript": "^5",
-    "vitest": "^3.1.3",
-    "@vitejs/plugin-react": "^4.4.1",
-    "vite-tsconfig-paths": "^5.1.4"
-  }
-}
