--- evals/evals/agent-051-cache-components-sync-io/PROMPT.md
@@ -1,1 +1,0 @@
-Enable Cache Components for this operations console. The navigation must stay immediately useful, while the request identifier, current time, and sampling value must still be generated per request. Finish with a production build and do not leave routes opted out.
