--- evals/evals/agent-051-cache-components-sync-io/app/layout.tsx
@@ -1,14 +1,0 @@
-import type { ReactNode } from 'react'
-import { RequestMetadata } from '@/components/request-metadata'
-
-export default function RootLayout({ children }: { children: ReactNode }) {
-  return (
-    <html lang="en">
-      <body>
-        <nav>Operations</nav>
-        <RequestMetadata />
-        {children}
-      </body>
-    </html>
-  )
-}
