--- evals/evals/agent-051-cache-components-sync-io/app/page.tsx
@@ -1,8 +1,0 @@
-export default function HomePage() {
-  return (
-    <main>
-      <h1>Deployment health</h1>
-      <p>All systems operational</p>
-    </main>
-  )
-}
