--- evals/evals/agent-051-cache-components-sync-io/next.config.ts
@@ -1,5 +1,0 @@
-import type { NextConfig } from 'next'
-
-const nextConfig: NextConfig = {}
-
-export default nextConfig
